logout 추가 & vue 분할
