+++ frontend/src/assets/css/main.css
@@ -1,0 +1,371 @@
+#r-sidebar {
+    width: 320px;
+    position: fixed;
+    top: 106px;
+    right: -320px;
+    height: 100vh;
+    z-index: 999;
+    background: #fff;
+    transition: all 0.3s;
+    border: 1px solid #e6e6e6;
+    box-shadow: 0 15px 0 0 rgba(0, 0, 0, 0.11), 0 5px 15px 0 rgba(0, 0, 0, 0.08);
+}
+#r-sidebar .r-header {
+    font-weight: 900;
+    background-color: #fafafa;
+}
+
+#r-sidebar .r-style {
+    padding: 12px 16px;
+    font-size: 18px;
+    line-height: 1.33334;
+    height: 61px;
+    display: flex;
+    align-items: center;
+    border-bottom: 1px solid #e6e6e6;
+}
+.s-coll-style {
+    padding: 12px 16px;
+    border-bottom: 1px solid #e6e6e6;
+}
+.s-coll-style > p {
+    margin-bottom: 5px;
+}
+.s-coll-style > li {
+    font-size: 14px;
+}
+
+.barRtoggle {
+    right: 0 !important;
+}
+
+/*CHAT MESSAGE*/
+.chat-message {
+    padding: 10px;
+    display: flex;
+}
+
+.chat-message .icon {
+    padding-right: 10px;
+}
+/*----------------------
+MainContents
+------------------------*/
+.mainwrapper {
+    height: calc(100vh - 105px);
+    overflow: hidden;
+}
+
+.h-inherit {
+    height: inherit;
+}
+/* .chat-wrapper {
+  height: 100%;
+  background: white;
+  border: 1px solid #e6e6e6;
+  width: 100%;
+} */
+
+/*chat contents wrapper*/
+.c-c-wrapper {
+    overflow: auto;
+    height: calc(100% - 141px);
+    padding: 20px 25px 0px 25px;
+    margin-bottom: 0px;
+}
+
+.c-i-wrapper {
+    height: 141px;
+    padding: 0px 25px;
+    display: flex;
+    /* border-top: 1px solid #e6e6e6; */
+}
+
+.nuj span {
+    font-size: 13px;
+    font-weight: bold;
+    margin-right: 3px;
+}
+
+.verti-align {
+    display: flex;
+    align-items: center;
+}
+
+/*-------------------------------------------
+   CUSTOM NAV STYLE
+---------------------------------------------*/
+.useridsty {
+    font-weight: 600;
+    margin-left: 5px;
+}
+
+.navbar-nav .nav-link {
+    padding: 0 0.5rem;
+}
+/*드롭다운 outline없애는데에 쓰임*/
+.nonoutline {
+    outline: none;
+}
+
+/*--------------------Scrollbar-----------------------------*/
+/* width */
+::-webkit-scrollbar {
+    width: 0.4em;
+}
+
+/* Track */
+::-webkit-scrollbar-track {
+    background: #f1f1f1;
+}
+
+/* Handle */
+::-webkit-scrollbar-thumb {
+    background: rgba(0, 0, 0, 0.2);
+    boarder-radius: 4px;
+}
+
+/* Handle on hover */
+::-webkit-scrollbar-thumb:hover {
+    background: #555;
+}
+
+/*
+    DEMO STYLE
+*/
+
+@import "https://fonts.googleapis.com/css?family=Poppins:300,400,500,600,700";
+body {
+    font-family: "Poppins", sans-serif;
+    background: #fafafa;
+}
+
+p {
+    font-family: "Poppins", sans-serif;
+    font-size: 1.1em;
+    font-weight: 300;
+    line-height: 1.7em;
+    color: #999;
+}
+
+a,
+a:hover,
+a:focus {
+    cursor: pointer;
+    color: inherit;
+    text-decoration: none;
+    transition: all 0.3s;
+}
+
+.navbar:nth-child(2) {
+    margin-top: 1px;
+    padding: 10px 15px;
+}
+
+.navbar {
+    /* position:fixed;
+    width:inherit; */
+
+    padding: 10px;
+    background: #fff;
+    border: none;
+    border-radius: 0;
+    box-shadow: 1px 1px 3px rgba(0, 0, 0, 0.1);
+}
+
+.navbar-btn {
+    box-shadow: none;
+    outline: none !important;
+    border: none;
+}
+
+.line {
+    width: 100%;
+    height: 1px;
+    border-bottom: 1px dashed #ddd;
+    margin: 40px 0;
+}
+
+.myflex-column {
+    display: flex;
+    flex-direction: column;
+}
+
+/* ---------------------------------------------------
+    SIDEBAR STYLE
+----------------------------------------------------- */
+.menulist {
+    flex-grow: 1;
+    overflow: auto;
+}
+
+.menulist-header-icon {
+    flex-grow: 1;
+    justify-content: flex-end;
+    display: flex;
+}
+
+.menulist-header-icon a {
+    display: flex;
+    text-decoration: none;
+    width: 17px;
+}
+.wrapper {
+    display: flex;
+    width: 100%;
+}
+
+#sidebar {
+    width: 250px;
+    position: fixed;
+    top: 0;
+    left: 0;
+    height: 100vh;
+    z-index: 999;
+    background: #7386D5;
+    color: #fff;
+    transition: all 0.3s;
+}
+
+#sidebar.active {
+    margin-left: -250px;
+}
+
+#sidebar .sidebar-header {
+    padding: 20px;
+    background: #6d7fcc;
+}
+
+#sidebar ul.components {
+    padding: 20px 0;
+    /* border-bottom: 1px solid #47748b; */
+    margin: 0px;
+}
+
+#sidebar ul .menulist-header:nth-child(1) {
+    padding: 0px 15px 0px 10px;
+}
+
+#sidebar ul .menulist-header {
+    color: #fff;
+    margin: 0px;
+    padding: 20px 15px 0px 10px;
+    font-weight: bold;
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+}
+
+#sidebar ul li a {
+    padding: 10px;
+    font-size: 15px;
+    display: block;
+}
+
+#sidebar ul li a:hover {
+    color: #7386D5;
+    background: #fff;
+}
+
+/* #sidebar ul li.active > a,
+a[aria-expanded="true"] {
+  color: #fff;
+} */
+
+#sidebar ul li a li {
+    overflow: hidden;
+    text-overflow: ellipsis;
+    margin-right: 30px;
+}
+
+.dropdown-toggle {
+    position: relative;
+}
+
+.dropdown-toggle::after {
+    display: block;
+    position: absolute;
+    top: 50%;
+    right: 20px;
+    transform: translateY(-50%);
+}
+
+ul ul a {
+    font-size: 0.9em !important;
+    padding-left: 30px !important;
+    background: #6d7fcc;
+}
+
+ul.CTAs {
+    padding: 20px;
+}
+
+ul.CTAs a {
+    text-align: center;
+    font-size: 0.9em !important;
+    display: block;
+    border-radius: 5px;
+    margin-bottom: 5px;
+}
+
+a.download {
+    background: #fff;
+    color: #7386D5;
+}
+
+a.article,
+a.article:hover {
+    background: #6d7fcc !important;
+    color: #fff !important;
+}
+
+/* ---------------------------------------------------
+    CONTENT STYLE
+----------------------------------------------------- */
+
+#m-wrapper {
+    word-break: break-all;
+    width: calc(100% - 250px);
+    min-height: 100vh;
+    transition: all 0.3s;
+    position: absolute;
+    top: 0;
+    right: 0;
+}
+
+#m-wrapper.active {
+    width: 100%;
+}
+/*-------------------------*/
+.navbar-nav .dropdown-menu {
+    position: absolute;
+}
+
+.navbar-nav {
+    flex-direction: row;
+}
+
+/* ---------------------------------------------------
+    MEDIAQUERIES
+----------------------------------------------------- */
+
+@media (max-width: 768px) {
+    #sidebar {
+        margin-left: -250px;
+    }
+    #sidebar.active {
+        margin-left: 0;
+    }
+    #m-wrapper {
+        width: 100%;
+    }
+    #m-wrapper.active {
+        width: calc(100% - 250px);
+    }
+    #sidebarCollapse span {
+        display: none;
+    }
+    .useridsty {
+        display: none;
+    }
+}
