+++ frontend/src/store/index.js
@@ -5,6 +5,7 @@ Vue.use(Vuex)
 
 export default new Vuex.Store({
   state: {
+    isLActive: false
   },
   mutations: {
   },
